Breweries & Beers: made changes to Jdbc, Controller and Dao as we did for Reviews so that they will show as soon as they are added

+++ java/src/main/java/com/techelevator/model/Brewery.java
@@ -2,7 +2,7 @@ package com.techelevator.model;
 
 public class Brewery {
    private Long breweryId;
-   private int userId;
+   private Long userId;
    private String name;
    private String websiteUrl;
    private String address;
@@ -22,10 +22,10 @@ public class Brewery {
       this.breweryId = breweryId;
    }
 
-   public int getUserId() {
+   public Long getUserId() {
       return userId;
    }
-   public void setUserId(int userId) {
+   public void setUserId(long userId) {
       this.userId = userId;
    }
